requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
packaging==25.0
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.4
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
websocket-client==1.9.0
wsproto==1.3.2
